requirements.txt
netbox>=4.1.0
django-filter>=21.1